Front-end Aesthetic Changes
Aesthetic changes to the current front end. Changed icons from square pictures to circular pngs so that they look better. Also changed a lot of constraints and dimensions so hopefully it looks better in different devices and orientations. Additionally changed contractions slidedeck components so that it can work like the website, as in: there isnt just one area to scroll through but a changing slidedeck depending on the top prompt.