Logic to Contractions Lessons
Logic to contractions lessons so that the images of the slidedecks change when the top arrows are pressed, as well as depending which button is touched in the contractions menu. All new images are included for the subsections except the last two because they are different on the assets folder.